+++ src/components/spartan/SInputOTP/styles.ts
@@ -1,13 +1,13 @@
 import { cva } from 'class-variance-authority';
 import { createBooleanVariation as cbv } from '@/helpers';
 
-export const inputOTPStyles = cva('relative flex cursor-text gap-1.5', {
+export const inputOtpStyles = cva('relative flex cursor-text gap-1.5', {
     variants: {
         disabled: cbv('opacity-50 pointer-events-none'),
     },
 });
 
-export const inputOTPItemStyles = cva(
+export const inputOtpItemStyles = cva(
     'flex h-12 w-[46px] shrink-0 items-center justify-center rounded-xl border bg-gray-100 border-gray-200',
     {
         variants: {
@@ -30,7 +30,7 @@ export const inputOTPItemStyles = cva(
     },
 );
 
-export const inputOTPItemTextStyles = cva('text-xl font-semibold text-gray-300', {
+export const inputOtpItemTextStyles = cva('text-xl font-semibold text-gray-300', {
     variants: {
         value: cbv('text-gray-900'),
         success: cbv('text-green-500'),