+++ src/stories/examples/AccountsDemo/components/stories.stories.ts
@@ -16,7 +16,7 @@ export const FormStory = createDefault({
     template: `<div class="h-full px-16 bg-white"><VForm v-bind="args" v-model="args.modelValue"/></div>`,
     transform: () => `<script setup lang="ts">
     import { ref } from 'vue';
-    import { SPlacetopayLogo, STab, STabItem, SInput, SInputBlock, SButton, SInputOTP, SInputOTPItem } from '@placetopay/spartan-vue';
+    import { SPlacetopayLogo, STab, STabItem, SInput, SInputBlock, SButton, SInputOtp, SInputOtpItem } from '@placetopay/spartan-vue';
     import { computed } from 'vue';
     import { ArrowLeftIcon } from '@heroicons/vue/20/solid';
     import { LockClosedIcon, KeyIcon } from '@heroicons/vue/24/outline';
@@ -104,15 +104,15 @@ export const FormStory = createDefault({
                     código de autenticación.
                 </p>
     
-                <SInputOTP v-model="otp" class="mb-6">
-                    <SInputOTPItem />
-                    <SInputOTPItem />
-                    <SInputOTPItem class="mr-2.5" />
+                <SInputOtp v-model="otp" class="mb-6">
+                    <SInputOtpItem />
+                    <SInputOtpItem />
+                    <SInputOtpItem class="mr-2.5" />
     
-                    <SInputOTPItem />
-                    <SInputOTPItem />
-                    <SInputOTPItem />
-                </SInputOTP>
+                    <SInputOtpItem />
+                    <SInputOtpItem />
+                    <SInputOtpItem />
+                </SInputOtp>
     
                 <div class="flex w-full">
                     <SButton class="mr-4 w-fit" variant="secondary" :left-icon="ArrowLeftIcon" @click="push('login')">