+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Changed
+- Renamed `SInputOTP` to `SInputOtp`.
+- Renamed `SInputOTPItem` to `SInputOtpItem`.
 - Adjusted `TBlockWrapperProps` types to be optional.
 
 ### Added